requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
guidata==2.2.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyparsing==3.0.9
PyQt6==6.3.0
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
python-dotenv==0.20.0
pytz==2022.1
QtPy==2.1.0
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
